add instagram user ID to models and DTOs and rename previous userId to forumUserId

+++ src/main/java/com/xhumanity/social/repository/TelegramUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.xhumanity.social.model.TelegramUser;
 
-public interface TelegramUserRepository extends JpaRepository<TelegramUser, Long> {
+public interface TelegramUserRepository extends JpaRepository<TelegramUser, Integer> {
 	
 	Optional<TelegramUser> findByUsername(@Param("username") final String username);
